==> SRC/notebooks/01_json_scraping.ipynb <==
import json


file_path = r"C:\Users\Admin\Desktop\DATAENGINEER\Cubix_data_engineer_vscode\data\Files\json_handling\spotify_playlist.json"

with open(file_path, "r") as json_file:
    data =json.load(json_file)
