==> SRC/notebooks/07_transform_load.ipynb <==
from datetime import datetime
from dateutil.relativedelta import relativedelta
import json
import os
import pandas as pd
pd.set_option("display.max_columns" , 30)
import requests

"""
1.get the data from s3
2.weather data transformations
3.taxi_trips transformations
4.update payment_type_master
5.update company_master
6.updatte taxi_trips with company and payment_type ids (replace the string values with ids from the latest amster table)
7.upload weather data to s3
8.upload taxi data to s3
9.upload the newest payment_type_master and company _master





"""

#### taxi_trips transformation codes

current_datetime = datetime.now() - relativedelta(months=2)
formatted_datetime= current_datetime.strftime("%Y-%m-%d")

url= f"https://data.cityofchicago.org/resource/ajtu-isnz.json?$where=trip_start_timestamp >= '{formatted_datetime}T00:00:00' AND trip_start_timestamp <='{formatted_datetime}T23:59:59'&$limit=30000"

headers= {"X-App-Token" : os.environ.get("CHICAGO_API_TOKEN")}
response= requests.get(url)
data= response.json()

taxi_trips =  pd.DataFrame(data)
#taxi_trips.head()

taxi_trips.drop(["pickup_census_tract","dropoff_census_tract"], axis=1, inplace=True)
taxi_trips.drop(["pickup_centroid_location","dropoff_centroid_location"], axis=1, inplace=True)

taxi_trips.dropna(inplace=True)

taxi_trips.rename(columns={"pickup_community_area":"pickup_community_area_id",
                  "dropoff_community_area":"dropoff_community_area_id"}, inplace=True)

taxi_trips["datetime_for_weather"]= pd.to_datetime(taxi_trips["trip_start_timestamp"])
taxi_trips["datetime_for_weather"]= taxi_trips["datetime_for_weather"].dt.floor("h")

taxi_trips.head()

### taxi trips transformation functions

def taxi_trips_transformations(taxi_trips:pd.DataFrame) ->pd.DataFrame:
    """Perform transformation with taxi data.

    Args:
        taxi_trips (pd.DataFrame): _description_
            The dataframe holding the daily taxi trips

    Returns:
        pd.DataFrame: _description_
            Cleaned transfromed dataframe holding the daily taxi trips.
    """
    if not isinstance(taxi_trips,pd.DataFrame):
        raise TypeError("taxi_trips is not a valid pandas Dataframe")
    
    
    
    taxi_trips.drop(["pickup_census_tract","dropoff_census_tract,dropoff_centroid_location,pickup_centroid_location"], axis=1, inplace=True)

    taxi_trips.dropna(inplace=True)

    taxi_trips.rename(columns={"pickup_community_area":"pickup_community_area_id",
    "dropoff_community_area":"dropoff_community_area_id"}, inplace=True)

    taxi_trips["datetime_for_weather"]= pd.to_datetime(taxi_trips["trip_start_timestamp"])
    taxi_trips["datetime_for_weather"]= taxi_trips["datetime_for_weather"].dt.floor("h")


    return taxi_trips

### company codes update

company_master = taxi_trips["company"].drop_duplicates().reset_index(drop=True)


company_master=pd.DataFrame(

{
 "company_id" : range(1,len(company_master)+1),
 "company": company_master




}
)
company_master

new_company_data = [
{"company":"Credit_card"},
{"company": "X"},
{"company": "Y"}
]
new_company_mapping = pd.DataFrame(new_company_data)
new_company_mapping

company_max_id=company_master["company_id"].max()

new_company_mapping["company"].values

new_companies_list = []

#for company in new_company_mapping["company"].values:
#    if company not in company_master["company"].values:
#        new_companies_list.append(company)
# one line
new_companies_list = [company for company in new_company_mapping["company"].values if company not in company_master["company"].values]

new_companies_df = pd.DataFrame({
"company_id":range(company_max_id+1, company_max_id+ len(new_companies_list)+1),
"company": new_companies_list


})
new_companies_df

updated_company_master = pd.concat([company_master, new_companies_df], ignore_index=True)
updated_company_master.tail()

def update_company_master(taxi_trips:pd.DataFrame,company_master:pd.DataFrame) -> pd.DataFrame:
    """Eytend the company master with companies if there are new companies

    Args:
        taxi_trips (pd.DataFrame): _description_
            Dataframe holding the daily taxi trips
        company_master (pd.DataFrame): _description_
            Dataframe holding the company master data


    Returns:
        pd.DataFrame: _description_
        The updated company master data , if new companies are in the taxi data they will be loaded to it
    """
    company_max_id=company_master["company_id"].max()
    new_companies_list = [company for company in taxi_trips["company"].values if company not in company_master["company"].values]
   
    new_companies_df = pd.DataFrame({
    "company_id":range(company_max_id+1, company_max_id+ len(new_companies_list)+1),
    "company": new_companies_list


})
    updated_company_master = pd.concat([company_master, new_companies_df], ignore_index=True)

    return updated_company_master

taxi_trips_company_only = pd.DataFrame({

    "company_id":[1,2,3],
    "company": ["4053 - 40193 Adwar H. Nikola","X","Y"]

})
taxi_trips_company_only

updated_company_master= update_company_master(taxi_trips_company_only, company_master)


### payment_type_master codes

payment_type_master = taxi_trips["payment_type"].drop_duplicates().reset_index(drop=True)


payment_type_master=pd.DataFrame(

{
 "payment_type_id" : range(1,len(payment_type_master)+1),
 "payment_type": payment_type_master
}
)
taxi_trips_payment_type_only = pd.DataFrame({

    "payment_type_id":[1,2,3],
    "payment_type": ["Credit Card","X","Y"]

})
taxi_trips_payment_type_only

def update_payment_type_master(taxi_trips:pd.DataFrame,payment_type_master:pd.DataFrame) -> pd.DataFrame:
    """Eytend the payment_type master with payment_type if there are new payment_types

    Args:
        taxi_trips (pd.DataFrame): _description_
            Dataframe holding the daily taxi trips
        payment_type (pd.DataFrame): _description_
            Dataframe holding the company master data


    Returns:
        pd.DataFrame: _description_
        The updated payment_type_master data , if new payment_types are in the taxi data they will be loaded to it
    """
    payment_type_max_id=payment_type_master["payment_type_id"].max()
    new_payment_types_list = [payment_type for payment_type in taxi_trips["payment_type"].values if payment_type not in payment_type_master["payment_type"].values]
   
    new_payment_type_df = pd.DataFrame({
    "payment_type_id":range(payment_type_max_id+1, payment_type_max_id+ len(new_payment_types_list)+1),
    "payment_type": new_payment_types_list


})
    updated_payment_type_master = pd.concat([payment_type_master, new_payment_type_df], ignore_index=True)

    return updated_payment_type_master

updated_payment_type_master= update_payment_type_master(taxi_trips_payment_type_only, payment_type_master)
updated_payment_type_master

### Creating a generic update master table function

def update_master( taxi_trips: pd.DataFrame,master: pd.DataFrame, id_column:str, value_column:str) -> pd.DataFrame:
    """Eytend the master Dataframe with new values if there are any aprameters.

    Args:
        taxi_trips (pd.DataFrame): _description_
            Dataframe holding the daily taxi trips
        master(pd.DataFrame): _description_
            Dataframe holding the master data
        id_column:str
            The id_column of the master Dataframe.
        value_column:str
            Name of the column in master_df containing the values

    Returns:
        pd.DataFrame: _description_
        The updated master data , if new values are in the taxi data they will be loaded to it
    """
    max_id=master[id_column].max()
    new_values_list = [value for value in taxi_trips[value_column].values if value not in master[value_column].values]
   
    new_values_df = pd.DataFrame({
    id_column: range(max_id+1, max_id+ len(new_values_list)+1),
    value_column: new_values_list


})
    updated_master = pd.concat([master, new_values_df], ignore_index=True)

    return updated_master

test_payment_type_master = update_master(taxi_trips_payment_type_only,payment_type_master,"payment_type_id","payment_type")
test_payment_type_master





test_company_master = update_master(taxi_trips_company_only,company_master,"company_id","company")
test_company_master.tail()

### update taxi trips with the most recent company_master and payment_type master codes

taxi_trips.head()

def update_taxi_trips_with_master_data(taxi_trips:pd.DataFrame,payment_type_master:pd.DataFrame,company_master: pd.DataFrame):
    """Update the taxi_trips dataframe with the company_master and payment_master ids, and delete the string columns.

    Args:
        taxi_trips (pd.DataFrame): _description_
            the dataframe with the daily taxi trips
        payment_type_master (pd.DataFrame): _description_
            The payment master table
        company_master (pd.DataFrame): _description_
            The company master table

    Returns:
        _type_: _description_
    """

    taxi_trips_id = taxi_trips.merge(payment_type_master,on="payment_type")
    taxi_trips_id = taxi_trips_id.merge(company_master,on="company")
    taxi_trips_id.drop(["payment_type","company"],axis=1, inplace=True)

    return taxi_trips_id

taxi_trips_id = update_taxi_trips_with_master_data(taxi_trips,payment_type_master,company_master)
taxi_trips_id

#### weather transformation function

def tansform_weather_data (weather_data: json) ->pd.DataFrame:
    """make daily transformation on the daily weather API response

    Args:
        weather_data (json): _description_
        Daily weather data from the Open Meteo API
    Returns:
        pd.DataFrame: _description_
            A Dataframe representation of the weather data
    """
    weather_data_filtered = {
        "datetime":weather_data ["hourly"]["time"],
        "temperature":weather_data ["hourly"]["temperature_2m"],
        "wind_speed":weather_data ["hourly"]["wind_speed_10m"],
        "rain":weather_data ["hourly"]["rain"],
        "precipitation":weather_data ["hourly"]["precipitation"],
}

    weather_df=pd.DataFrame(weather_data_filtered)
    weather_df["datetime"] = pd.to_datetime(weather_df["datetime"])

    return weather_df

current_datetime = datetime.now() - relativedelta(months=2)
formatted_datetime= current_datetime.strftime("%Y-%m-%d")

url= "https://archive-api.open-meteo.com/v1/era5"



params={
    "latitude" :41.85,
    "longitude" :-87.68,
    "start_date" :formatted_datetime,
    "end_date" : formatted_datetime,
    "hourly" : "temperature_2m,wind_speed_10m,precipitation,rain"   
     }
response= requests.get(url,params=params)
weather_data = response.json()

weather_data_df = tansform_weather_data(weather_data)
