==> SRC/notebooks/09_HF2.ipynb <==
## Create a function for reading json from a s3 bucket

from io import StringIO
import json
import boto3
import pandas as pd

def read_json_from_s3(bucket: str, key: str) -> dict:
    """
    Reads a JSON file from an S3 bucket and returns the JSON data.

    Parameters:
        bucket : The name of the S3 bucket.
        key : The path of the file in the S3 bucket.

    Returns:
        dict: The  return is a python dictionary.
    """
    s3 = boto3.client('s3')
    response = s3.get_object(Bucket=bucket, Key=key)
    content = response["Body"]
    return json.loads(content.read())

def lambda_handler(event, context):
    s3=boto3.client("s3") 

    bucket="dataengineerea"

    raw_weather_folder="raw_data/to_be_processed/weather_data/"
    row_taxi_trips_folder="raw_data/to_be_processed/taxi_data/"
    target_taxi_trips_folder="raw_data/processed/taxi_data/"
    target_weather_folder="raw_data/processed/weather _data/"

    transformed_taxi_trips_folder="transformed_data/taxi_trips/"
    transformed_weather_folder="transformed_data/weather/"
    
    payment_type_master_folder = "transformed_data/payment_type/"
    company_master_folder = "transformed_data/company/"
    
    payment_type_master_filename = "payment_type_master.csv"
    company_master_filename = "company_master.csv"
    

    payment_type_master = read_csv_from_s3(bucket=bucket, path=payment_type_master_folder, filename=payment_type_master_filename)
    company_master = read_csv_from_s3(bucket=bucket, path=company_master_folder, filename=company_master_filename)
        
    for file in s3.list_objects(Bucket=bucket, Prefix=row_taxi_trips_folder)["Contents"]:taxi_trip_key = file['Key']

        if taxi_trip_key.split("/")[-1].strip() != "":
            if taxi_trip_key.split(".")[1] == "json":
                filename = taxi_trip_key.split("/")[-1]

                # Use the new function to read JSON data for the taxi_trips
                taxi_trip_data_json = read_json_from_s3(bucket, taxi_trip_key)
                
    
    
    for file in s3.list_objects(Bucket=bucket, Prefix=raw_weather_folder)["Contents"]:weather_key = file["Key"]
        if weather_key.endswith(".json"):
            print(f"Processing file: {weather_key}")
            
            filename = weather_key.split("/")[-1]
            
            # Use the new function to read JSON data for weather
            weather_data_json = read_json_from_s3(bucket=bucket, key=weather_key)
