==> SRC/notebooks/10_SQL_queries_HF3.ipynb <==
## Which 10 companies make the most money? Use two columns: Company name and the sum of the fare. Order by descending order.Make sure that the sum of the fares (total_fare) is not in scientific notation

SELECT 
    c.company AS company_name,
    CAST(SUM(t.fare) AS DECIMAL(18, 2)) AS total_fare
FROM 
    "AwsDataCatalog"."chicago_taxi_db"."taxi_trips" t
JOIN 
    "AwsDataCatalog"."chicago_taxi_db"."company" c
ON 
    t.company_id = c.company_id
GROUP BY 
    c.company
ORDER BY 
    total_fare DESC
LIMIT 10;

## Show the 10 pickup community areas with the most rides. Use two columns: community area name and count of rides per area, in descending order for the rides.

SELECT 
    ca.community_name AS community_area_name, 
    COUNT(t.trip_id) AS ride_count
FROM 
    chicago_taxi_db.taxi_trips t
JOIN 
    chicago_taxi_db.community_areas ca
ON 
    t.pickup_community_area_id = ca.area_code
GROUP BY 
    ca.community_name
ORDER BY 
    ride_count DESC
LIMIT 10;

## Get the count of taxi rides per day of week. In other words, we'd like to see which day has the most rides. Use the names of the days instead of numbers (Monday = 1, Tuesday = 2, etc.).HINT: Check CASE WHEN statements for replacing names of days. For ordering the results by the day you can use MIN(day_of_the_week)

SELECT
    d.dayofweek,      
    COUNT(*) AS total_rides
FROM taxi_trips t
JOIN date d
    ON substring(t.trip_start_timestamp, 1, 10) = d.date
GROUP BY
    d.dayofweek       
ORDER BY
    MIN(d.dayofweek);
